+++ Enigma/Enigma.Client/Models/ProfileResponse.cs
@@ -20,7 +20,7 @@ public class ProfileUserData
     public string Username { get; set; } = string.Empty;
 
     [JsonPropertyName("maze_nuggets")]
-    public int MazeNuggets { get; set; }
+    public long MazeNuggets { get; set; }
 
     [JsonPropertyName("friends")]
     public List<string> Friends { get; set; } = [];
